Trait tikv_util::codec::number::NumberEncoder

pub trait NumberEncoder: Write {
    fn encode_i64(&mut self, v: i64) -> Result<()> { ... }

    fn encode_i64_desc(&mut self, v: i64) -> Result<()> { ... }

    fn encode_u64(&mut self, v: u64) -> Result<()> { ... }

    fn encode_u64_desc(&mut self, v: u64) -> Result<()> { ... }

    fn encode_u32(&mut self, v: u32) -> Result<()> { ... }

    fn encode_u16(&mut self, v: u16) -> Result<()> { ... }

    fn encode_var_i64(&mut self, v: i64) -> Result<()> { ... }

    fn encode_var_u64(&mut self, mut v: u64) -> Result<()> { ... }

    fn encode_f64(&mut self, f: f64) -> Result<()> { ... }

    fn encode_f64_desc(&mut self, f: f64) -> Result<()> { ... }

    fn encode_u16_le(&mut self, v: u16) -> Result<()> { ... }

    fn encode_f32_le(&mut self, v: f32) -> Result<()> { ... }

    fn encode_u32_le(&mut self, v: u32) -> Result<()> { ... }

    fn encode_i32_le(&mut self, v: i32) -> Result<()> { ... }

    fn encode_f64_le(&mut self, v: f64) -> Result<()> { ... }

    fn encode_i64_le(&mut self, v: i64) -> Result<()> { ... }

    fn encode_u64_le(&mut self, v: u64) -> Result<()> { ... }
}

Provided methods

fn encode_i64(&mut self, v: i64) -> Result<()>

Writes the encoded value to buf. It guarantees that the encoded value is in ascending order for comparison.

fn encode_i64_desc(&mut self, v: i64) -> Result<()>

Writes the encoded value to buf. It guarantees that the encoded value is in descending order for comparison.

fn encode_u64(&mut self, v: u64) -> Result<()>

Writes the encoded value to slice buf. It guarantees that the encoded value is in ascending order for comparison.

fn encode_u64_desc(&mut self, v: u64) -> Result<()>

Writes the encoded value to slice buf. It guarantees that the encoded value is in descending order for comparison.

fn encode_u32(&mut self, v: u32) -> Result<()>

Writes the encoded value to slice buf in big endian order.

fn encode_u16(&mut self, v: u16) -> Result<()>

Writes the encoded value to slice buf in big endian order.

fn encode_var_i64(&mut self, v: i64) -> Result<()>

Writes the encoded value to slice buf. Note that the encoded result is not memcomparable.

fn encode_var_u64(&mut self, mut v: u64) -> Result<()>

Writes the encoded value to slice buf. Note that the encoded result is not memcomparable.

fn encode_f64(&mut self, f: f64) -> Result<()>

Writes the encoded value to slice buf. It guarantees that the encoded value is in ascending order for comparison.

fn encode_f64_desc(&mut self, f: f64) -> Result<()>

Writes the encoded value to slice buf. It guarantees that the encoded value is in descending order for comparison.

fn encode_u16_le(&mut self, v: u16) -> Result<()>

Writes u16 numbers in little endian order.

fn encode_f32_le(&mut self, v: f32) -> Result<()>

Writes f32 numbers in little endian order.

fn encode_u32_le(&mut self, v: u32) -> Result<()>

Writes u32 numbers in little endian order.

fn encode_i32_le(&mut self, v: i32) -> Result<()>

Writes i32 numbers in little endian order.

fn encode_f64_le(&mut self, v: f64) -> Result<()>

Writes f64 numbers in little endian order.

fn encode_i64_le(&mut self, v: i64) -> Result<()>

Writes i64 numbers in little endian order.

fn encode_u64_le(&mut self, v: u64) -> Result<()>

Writes u64 numbers in little endian order.

Loading content...