Module tikv_util::config

Modules

check_data_dir

check_kernel

Structs

OptionReadableSize
This trivial type is needed, because we can't define the From<Option<ReadableSize>> and Into<Option<ReadableSize>> trait for ConfigValue which is needed to derive Configuration trait for BlockCacheConfig

ReadableDuration

ReadableSize

TomlWriter
TomlWriter use to update the config file and only cover the most commom toml format that used by tikv config file, toml format like: quoted keys, multi-line value, inline table, etc, are not supported, see https://github.com/toml-lang/toml for more detail.

Tracker

VersionTrack

Enums

ConfigError

LogFormat

TomlLine
TomlLine use to parse one line content of a toml file

Constants

DATA_MAGNITUDE

DAY

GB

HOUR

KB

MB

MINUTE

MS

PB

SECOND

TB

TIME_MAGNITUDE_1

TIME_MAGNITUDE_2

TIME_MAGNITUDE_3

UNIT

Functions

canonicalize_log_dir

canonicalize_path

canonicalize_sub_path

check_addr
check_addr validates an address. Addresses are formed like "Host:Port". More details about Host and Port can be found in WHATWG URL Standard.

check_data_dir

check_data_dir_empty

check_kernel
check_kernel_params checks kernel parameters, following are checked so far:

check_max_open_fds

ensure_dir_exist

get_file_count