Module tikv_util::file

Structs

Sha256Reader
Wrapper of a reader which computes its SHA-256 hash while reading.

Constants

DIGEST_BUFFER_SIZE

Functions

calc_crc32
Calculates the given file's Crc32 checksum.

calc_crc32_and_size
Calculates crc32 and decrypted size for a given reader.

calc_crc32_bytes
Calculates the given content's CRC32 checksum.

create_dir_if_not_exist
Creates a new, empty directory at the provided path. Returns true on success, false if the directory already exists. Otherwise the raw error will be returned.

delete_dir_if_exist
Deletes given path from file system. Returns true on success, false if the directory doesn't exist. Otherwise the raw error will be returned.

delete_file_if_exist
Deletes given path from file system. Returns true on success, false if the file doesn't exist. Otherwise the raw error will be returned.

file_exists

get_file_size

sha256

sync_dir
Call fsync on directory by its path