Module tikv_util::future

Structs

BatchCommandsWaker

Functions

create_stream_with_buffer
Create a stream proxy with buffer representing the remote stream. The returned task will receive messages from the remote stream as much as possible.

paired_future_callback
Generates a paired future and callback so that when callback is being called, its result is automatically passed as a future result.

paired_must_called_future_callback

poll_future_notify
Polls the provided future immediately. If the future is not ready, it will register the waker. When the event is ready, the waker will be notified, then the internal future is immediately polled in the thread calling wake().