Module tikv_util::future_pool

This mod implemented a wrapped future pool that supports on_tick() which is invoked no less than the specific interval.

Modules

builder

metrics

Structs

Builder

Config

Env

Full

FuturePool

FuturePoolRunner

Constants

THREAD_LAST_TICK_TIME

TICK_INTERVAL

Functions

try_tick_thread
Tries to trigger a tick in current thread.

Type Definitions

ThreadPool