Module tikv_util::logger

Modules

file_log

formatter

Structs

LogAndFuse

LogCost
Special struct for slow log cost serializing

LogDispatcher
Dispatches logs to a normal Drain or a slow-log specialized Drain by tag

RocksFormat

Serializer

SlowCostSerializer

SlowLogFilter

TikvFormat

Enums

Level
Logging level associated with a logging Record

Constants

DATETIME_ROTATE_SUFFIX

SLOG_CHANNEL_OVERFLOW_STRATEGY

SLOG_CHANNEL_SIZE

TIMESTAMP_FORMAT

Statics

LOG_LEVEL

Traits

FilterFn
Function that can be used in Filter drain

Functions

convert_log_level_to_slog_level

convert_slog_level_to_log_level

file_writer
Constructs a new file writer which outputs log to a file at the specified path. The file writer rotates for the specified timespan.

get_level_by_string

get_log_level

get_string_by_level

get_unified_log_level

init_log

json_format
Formats output logs to JSON format.

rocks_text_format
Same as text_format, but is adjusted to be closer to vanilla RocksDB logger format.

set_global_logger

set_log_level

term_writer
Constructs a new terminal writer which outputs logs to stderr.

text_format
Formats output logs to "TiDB Log Format".

write_log_fields
Writes log fields to decorator. See log-fields

write_log_header
Writes log header to decorator. See log-header

write_log_msg
Writes log message to decorator. See log-message