Struct tikv_util::logger::file_log::RotatingFileLogger

pub struct RotatingFileLogger {
    path: PathBuf,
    file: File,
    rename: Box<dyn Send + Fn(&Path) -> Result<PathBuf>>,
    rotators: Vec<Box<dyn Rotator>>,
}

This FileLogger will iterate over a series of Rotators, once the context trigger the Rotator, it will execute a rotation.

After rotating, the original log file would be renamed to “{original name}.{%Y-%m-%d-%H:%M:%S}”. Note: log file will not be compressed or otherwise modified.

Fields
path: PathBuffile: Filerename: Box<dyn Send + Fn(&Path) -> Result<PathBuf>>rotators: Vec<Box<dyn Rotator>>