Function tikv_util::logger::formatter::write_escaped_str

pub fn write_escaped_str<W: ?Sized>(writer: &mut W, value: &str) -> Result<()> where
    W: Write, 

According to RFC: Unified Log Format, escapes the given data and writes it into a writer. If there is no character need json encode, it writes the data into the writer directly. Else, it serializes the given data structure as JSON into a writer.