Module tikv_util::metrics

Re-exports

pub use self::allocator_metrics::monitor_allocator_stats;

Modules

allocator_metrics

metrics_reader

process_linux
This module is a subset of rust-prometheus’s process collector, without the fd collector to avoid memory fragmentation issues when open fd is large.

threads_linux

Structs

CRITICAL_ERROR

HistogramReader

ThreadInfoStatistics
Use to collect cpu usages and disk I/O rates

Functions

cpu_total

dump

get_thread_ids
Gets thread ids of the given process id. WARN: Don’t call this function frequently. Otherwise there will be a lot of memory fragments.

monitor_process
Monitors current process.

monitor_threads
Monitors threads of the current process.