Module tikv_util::metrics::process_linux

This module is a subset of rust-prometheus's process collector, without the fd collector to avoid memory fragmentation issues when open fd is large.

Structs

BOOT_TIME

CLK_TCK

PAGESIZE

ProcessCollector
A collector to collect process metrics.

Functions

monitor_process
Monitors current process.