Module tikv_util::metrics::threads_linux

Structs

CLK_TCK

Metrics

ThreadInfoStatistics
Use to collect cpu usages and disk I/O rates

ThreadMetrics

ThreadsCollector
A collector to collect threads metrics, including CPU usage and threads state.

TidRetriever
A helper that buffers the thread id list internally.

Constants

TID_MAX_UPDATE_INTERVAL

TID_MIN_UPDATE_INTERVAL

Functions

collect_metrics_by_name

cpu_total

get_name

get_thread_ids
Gets thread ids of the given process id. WARN: Don't call this function frequently. Otherwise there will be a lot of memory fragments.

monitor_threads
Monitors threads of the current process.

sanitize_thread_name
Sanitizes the thread name. Keeps a-zA-Z0-9_:, replaces - and with _, and drops the others.

state_to_str

to_io_err

update_metric