Module tikv_util::mpsc

This module provides an implementation of mpsc channel based on crossbeam_channel. Comparing to the crossbeam_channel, this implementation supports closed detection and try operations.

Modules

batch

Structs

LooseBoundedSender
A sender of channel that limits the maximun pending messages count loosely.

Receiver
The receive end of a channel.

Sender
A sender that can be closed.

State

Constants

CHECK_INTERVAL

Functions

bounded
Create a bounded channel.

loose_bounded
Create a loosely bounded channel with the given capacity.

unbounded
Create an unbounded channel.