Module tikv_util::mpsc::batch

Structs

BatchReceiver
BatchReceiver is a futures::Stream, which returns a batched type.

Notifier
Notifier is used to notify receiver whenever you want.

Receiver

Sender

State

VecCollector

Traits

BatchCollector
A Collector Used in BatchReceiver.

Functions

bounded
Creates a bounded channel with a given notify_size, which means if there are more pending messages in the channel than notify_size, the Sender will auto notify the Receiver.

unbounded
Creates a unbounded channel with a given notify_size, which means if there are more pending messages in the channel than notify_size, the Sender will auto notify the Receiver.