Module tikv_util::threadpool

Structs

DefaultContext

DefaultContextFactory

FifoQueue

ScheduleState

Scheduler

Task

ThreadPool
ThreadPool is used to execute tasks in parallel. Each task would be pushed into the pool, and when a thread is ready to process a task, it will get a task from the pool according to the ScheduleQueue provided in initialization.

ThreadPoolBuilder

Worker

Constants

DEFAULT_QUEUE_CAPACITY

DEFAULT_TASKS_PER_TICK

DEFAULT_THREAD_COUNT

NAP_SECS

QUEUE_MAX_CAPACITY

Traits

Context

ContextFactory