Module tikv_util::time

Modules

inner

Structs

CoarseClock
A coarse clock for async_speed_limit.

Duration
A Duration type to represent a span of time, typically used for system timeouts.

Monitor

SlowTimer

ThreadReadId
ReadId to judge whether the read requests come from the same GRPC stream.

UnixSecs
A time in seconds since the start of the Unix epoch.

Enums

Instant
A measurement of a monotonically increasing clock. It's similar and meant to replace std::time::Instant, for providing extra features.

Constants

DEFAULT_SLOW_SECS

DEFAULT_WAIT_MS

MILLISECOND_PER_SECOND

NANOSECONDS_PER_MILLISECOND

NANOSECONDS_PER_SECOND

READ_SEQUENCE

Functions

duration_to_ms
Converts Duration to milliseconds.

duration_to_nanos
Converts Duration to nanoseconds.

duration_to_sec
Converts Duration to seconds.

monotonic_now

monotonic_raw_now

Type Definitions

Limiter
A limiter which uses the coarse clock for measurement.