Module tikv_util::yatp_pool

Modules

future_pool
This mod implemented a wrapped future pool that supports on_tick() which is invoked no less than the specific interval.

metrics

Structs

Config

DefaultTicker

Full

FuturePool

TickerWrapper

YatpPoolBuilder

YatpPoolRunner

Constants

TICK_INTERVAL

Traits

PoolTicker