Function tikv_util::future::create_stream_with_buffer

pub fn create_stream_with_buffer<T, S>(
    s: S, 
    size: usize
) -> (impl Stream<Item = T> + Send + 'static, impl Future<Output = ()> + Send + 'static) where
    S: Stream<Item = T> + Send + 'static,
    T: Send + 'static, 

Create a stream proxy with buffer representing the remote stream. The returned task will receive messages from the remote stream as much as possible.