Module tikv_util::worker

Worker provides a mechanism to run tasks asynchronously (i.e. in the background) with some additional features, for example, ticks.

A worker contains:

A runner (which should implement the Runnable trait): to run tasks one by one or in batch.

A scheduler: to send tasks to the runner, returns immediately.

Briefly speaking, this is a mpsc (multiple-producer-single-consumer) model.

Modules

future

metrics

Structs

Builder

DefaultRunnerWithTimer

FutureScheduler
Scheduler provides interface to schedule task to underlying workers.

FutureWorker
A worker that can schedule time consuming tasks.

Scheduler
Scheduler provides interface to schedule task to underlying workers.

Stopped

Worker
A worker that can schedule time consuming tasks.

Enums

ScheduleError

Traits

FutureRunnable

Runnable

RunnableWithTimer

Functions

dummy_future_scheduler

dummy_scheduler
Creates a scheduler that can't schedule any task.

fill_task_batch
Fills buffer with next task batch coming from rx.

poll